Start versioning from 0.3.0
This requires moving every current translation into a 0.3.0/ directory and doing
some default redirection when people hit keepachangelog.com they should
always be redirected to the latest version.

Currently that's hardcoded to 0.3.0 but we're going to surely use something akin
to a version.rb to define that and use it across the codebase.