Greatly simplify default version redirection
Can’t believe I didn’t think of it sooner but obviously we might
as well check all available version for each language and redirect
by default to the latest possible.

Now I’ll have to figure out a way to display when the latest
available version *for a language* isn’t the latest version of
the manifesto (so if it doesn’t match the latest English version).